split out demodulator class from applications/gqrx/receiver* files, rename a few things to be consistent
